+++ packages/cosmwasm/src/cosmwasmclient.ts
@@ -13,6 +13,7 @@ import {
   PubKey,
   setupAuthExtension,
   StdTx,
+  uint64ToNumber,
 } from "@cosmjs/sdk38";
 
 import { setupWasmExtension, WasmExtension } from "./lcdapi/wasm";
@@ -234,8 +235,8 @@ export class CosmWasmClient {
         address: value.address,
         balance: value.coins,
         pubkey: value.public_key ? decodeBech32Pubkey(value.public_key) : undefined,
-        accountNumber: value.account_number,
-        sequence: value.sequence,
+        accountNumber: uint64ToNumber(value.account_number),
+        sequence: uint64ToNumber(value.sequence),
       };
     }
   }