+++ packages/sdk38/src/cosmosclient.ts
@@ -3,7 +3,7 @@ import { fromBase64, fromHex, toHex } from "@cosmjs/encoding";
 import { Uint53 } from "@cosmjs/math";
 
 import { Coin } from "./coins";
-import { AuthExtension, BroadcastMode, LcdClient, setupAuthExtension } from "./lcdapi";
+import { AuthExtension, BroadcastMode, LcdClient, setupAuthExtension, uint64ToNumber } from "./lcdapi";
 import { Log, parseLogs } from "./logs";
 import { decodeBech32Pubkey } from "./pubkey";
 import { CosmosSdkTx, PubKey, StdTx } from "./types";
@@ -235,8 +235,8 @@ export class CosmosClient {
         address: value.address,
         balance: value.coins,
         pubkey: value.public_key ? decodeBech32Pubkey(value.public_key) : undefined,
-        accountNumber: value.account_number,
-        sequence: value.sequence,
+        accountNumber: uint64ToNumber(value.account_number),
+        sequence: uint64ToNumber(value.sequence),
       };
     }
   }